Break the blog out into a blueprint
This is to help a bit with some micropub work.
