+++ komorebi/templates/archive.html
@@ -31,7 +31,7 @@
 	{%- if entry.n == 0 %}
 		<dd class="empty">&mdash;</dd>
 	{%- else %}
-		<dd><a href="{{ url_for('month', year=entry.year, month='{:02}'.format(entry.month)) }}">{{ entry.n }}</a></dd>
+		<dd><a href="{{ url_for('.month', year=entry.year, month='{:02}'.format(entry.month)) }}">{{ entry.n }}</a></dd>
 	{%- endif %}
 {%- endfor %}
 </dl>
