+++ komorebi/templates/base.html
@@ -6,7 +6,7 @@
 		<link rel="preconnect" href="https://www.youtube.com/" crossorigin>
 		<meta name="Author" content="Keith Gaughan">
 		<meta name="Copyright" content="Copyright (c) Keith Gaughan, 2001-2019">
-		<link title="Feed" rel="alternate" type="application/atom+xml" href="{{ url_for('feed') }}">
+		<link title="Feed" rel="alternate" type="application/atom+xml" href="{{ url_for('blog.feed') }}">
 		<link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Inconsolata:400,700|Raleway:400,400i,600&display=swap">
 		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/8.0.0/normalize.min.css" integrity="sha256-oSrCnRYXvHG31SBifqP2PM1uje7SJUyX0nTwO2RJV54=" crossorigin="anonymous">
 		<link rel="stylesheet" media="screen" href="{{ url_for('static', filename='screen.css') }}">
@@ -17,13 +17,13 @@
 		<header>
 			<h1>Inklings: a tumblelog</h1>
 			<nav>
-				<a href="{{ url_for('latest') }}">Home</a> &#x2055;
-				<a href="{{ url_for('archive') }}">Archive</a> &#x2055;
-				<a href="{{ url_for('add_entry') }}">Add</a> &#x2055;
+				<a href="{{ url_for('blog.latest') }}">Home</a> &#x2055;
+				<a href="{{ url_for('blog.archive') }}">Archive</a> &#x2055;
+				<a href="{{ url_for('blog.add_entry') }}">Add</a> &#x2055;
 				<a href="https://keith.gaughan.ie/about/">About Me</a> &#x2055;
 				<a href="https://keith.gaughan.ie/">Weblog</a> &#x2055;
 				<a href="https://github.com/kgaughan">Github</a> &#x2055;
-				<a rel="alternate" type="application/atom+xml" href="{{ url_for('feed') }}">Feed</a>
+				<a rel="alternate" type="application/atom+xml" href="{{ url_for('blog.feed') }}">Feed</a>
 			</nav>
 		</header>
 		<main>{% block content %}{% endblock %}</main>
