+++ komorebi/templates/macros.html
@@ -2,7 +2,7 @@
 <article>
 
 	<header>
-		<div class="permalink"><a title="Permalink" href="{{ url_for('entry', entry_id=entry.id) }}">&infin;</a></div>
+		<div class="permalink"><a title="Permalink" href="{{ url_for('.entry', entry_id=entry.id) }}">&infin;</a></div>
 
 		{% if entry.link %}
 			<h2><a href="{{ entry.link }}">{{ entry.title }}</a></h2>
@@ -10,14 +10,14 @@
 			<h2>{{ entry.title }}</h2>
 		{% endif -%}
 		<div class="metadata">
-		<span title="Published">&#128337; <time datetime="{{ entry.time_c | iso_date }}">{{ entry.time_c | iso_date }}</time></span>
+		<span title="Published">&#128337; <time datetime="{{ entry.time_c | to_iso_date }}">{{ entry.time_c | to_iso_date }}</time></span>
 		{%- if entry.via -%}
 		; via <a href="{{ entry.via }}">{{ entry.via | extract_hostname }}</a>
 		{%- endif -%}
 		{%- if entry.link -%}
 		; try the <a rel="nofollow" href="https://web.archive.org/web/*/{{ entry.link }}">Wayback Machine</a>
 		{% endif %}
-		| <a rel="nofollow" href="{{ url_for('edit_entry', entry_id=entry.id) }}">edit</a>
+		| <a rel="nofollow" href="{{ url_for('.edit_entry', entry_id=entry.id) }}">edit</a>
 		</div>
 	</header>
